DBのEntityをdata層直下に移動
